- Findbugs fixes.

+++ src/com/markupartist/iglaset/activity/DrinkDetailActivity.java
@@ -497,7 +497,7 @@ public class DrinkDetailActivity extends ListActivity {
                 ratingValue.setText("Ditt betyg " + userRating);
                 final RatingBar ratingBar = (RatingBar) layout.findViewById(R.id.add_user_rating);
                 ratingBar.setNumStars(5);
-                ratingBar.setStepSize((float) 0.5);
+                ratingBar.setStepSize(0.5f);
                 ratingBar.setRating(userRating / 2);
                 ratingBar.setOnRatingBarChangeListener(new OnRatingBarChangeListener() {
                     public void onRatingChanged(RatingBar ratingBar, float rating,
